backend/requirements.txt
google-genai
fastapi
websockets
uvicorn